Reverted breaking change from commit 137d16c. For whatever reason, this particular change broke the `writeDexFile()` function (see issues #59 and #63 , for example)
I don't really know why switching from `FluentIterable` to `TransformedIterator` broke things (I haven't really been digging on that), but it seems to be the root cause.
All tests were passing (as they do with this new commit), but the generated .dex was corrupted and not runnable.
Now, a simple example with `DexFileFactory.loadDexFile()` and `DexFileFactory.writeDexFile()` does generate a valid, runnable .dex file.